deployop update

+++ .github/scripts/validate-deploy.sh
@@ -8,9 +8,9 @@ NAMESPACE=$(cat .namespace)
 BRANCH="main"
 SERVER_NAME="default"
 TYPE="base"
-LAYER="3-applications"
+LAYER="2-services"
 
-COMPONENT_NAME="turbonomic"
+COMPONENT_NAME="turbo"
 
 mkdir -p .testrepo
 
@@ -54,6 +54,19 @@ fi
 
 ### add more validation logic when done ####
 #kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
+
+count=0
+until kubectl rollout status deployment/t8c-operator -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
+  echo "Waiting for turbo operator rollout to deploy"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for turbo operator to deploy"
+  exit 1
+fi
+
 
 cd ..
 rm -rf .testrepo
